+++ src/sources/acled.py
@@ -10,8 +10,8 @@ from typing import ClassVar
 import numpy as np
 import pandas as pd
 
+from _fb_types import SourceType
 from _schemas import AcledResolutionFrame
-from _types import SourceType
 
 from ._dataset import DatasetSource
